requirements.txt
pandas
fastapi
pydantic
uvicorn
python-dotenv
sqlalchemy
pymysql
